+++ Web/TimeBox.Web/Controllers/PlannedTasksController.cs
@@ -1,5 +1,6 @@
 ﻿namespace TimeBox.Web.Controllers
 {
+    using System;
     using System.Threading.Tasks;
 
     using Microsoft.AspNetCore.Authorization;
@@ -48,7 +49,17 @@
             }
 
             var user = await this.userManager.GetUserAsync(this.User);
-            await this.plannedTasksService.CreateAsync(input, user.Id);
+
+            try
+            {
+                await this.plannedTasksService.CreateAsync(input, user.Id);
+            }
+            catch (Exception ex)
+            {
+                this.ModelState.AddModelError(string.Empty, ex.Message);
+                input.CategoriesItems = this.categoriesService.GetAllAsKeyValuePairs();
+                return this.View(input);
+            }
 
             return this.Redirect("/PlannedTasks/Schedule");
         }
@@ -69,6 +80,12 @@
         {
             var user = await this.userManager.GetUserAsync(this.User);
             var plannedTask = this.plannedTasksService.ById(user, id);
+
+            if (this.ModelBinderFactory == null)
+            {
+                return this.RedirectToAction("NotFoundError", "Error");
+            }
+
             return this.View(plannedTask);
         }
 
@@ -92,6 +109,12 @@
             }
 
             await this.plannedTasksService.UpdateAsync(id, input);
+
+            if (input == null)
+            {
+                return this.RedirectToAction("NotFoundError", "Error");
+            }
+
             return this.Redirect("/PlannedTasks/Schedule");
         }
 
@@ -100,6 +123,11 @@
 
         public async Task<IActionResult> Delete(int id)
         {
+            if (!this.plannedTasksService.Exists(id))
+            {
+                return this.RedirectToAction("NotFoundError", "Error");
+            }
+
             await this.plannedTasksService.DeleteAsync(id);
             return this.Redirect("/PlannedTasks/Schedule");
         }
